// components/Utils/enums.tsx
export enum Difficulty {
    Hard = "HARD",
    Medium = "MEDIUM",
    Low = "LOW",
  }

export enum MarketPlace {
  Key = "99BydNFLlCzTLNnb51Rp"
}

  export enum Status {
    Active = "ACTIVE",
    Progress = "PROGRESS",
    Completed = "COMPLETED"
  }

  export enum ModalType {
    Success = "bg-primary-green",
    Error = "bg-primary-red",
    Dark = "bg-primary-black",
    Default = "bg-slate-50"
  }

  export enum Ranks {
    Newbie = "Newbie",
    Iron = "Iron",
    Bronze = "Bronze",
    Silver = "Silver",
    Gold = "Gold",
    Platinum = "Platinum",
    Diamond = "Diamond",
    Master = "Master",
    Grandmaster = "Grandmaster",
    Challenger = "Challenger"
  }